Crate gloo_file

Working with files and blobs on the Web.

These APIs come in two flavors:

a callback style (that more directly mimics the JavaScript APIs), and

a Future API.

Modules

callbacks

Structs

Blob

A Blob.

File

A File.

FileList

A list of files, for example from an <input type="file">.

ObjectUrl

A resource wrapper around URL.createObjectURL / URL.revokeObjectURL.

Enums

FileReadError

Traits

BlobContents

This trait is used to overload the Blob::new_with_options function, allowing a variety of types to be used to create a Blob. Ignore this, and use &[u8], &str, etc to create a Blob.