Module gloo_file::callbacks

Structs

FileReader

A guard object that aborts the file read when dropped (if the read isn’t already finished).

Functions

read_as_array_buffer

Asynchronously converts the blob into an array buffer and then passes it to the callback.

read_as_bytes

Asynchronously converts the blob into a Vec<u8> and then passes it to the callback.

read_as_data_url

Asynchronously converts the blob into a base64 encoded data: URL and then passes it to the callback.

read_as_text

Asynchronously converts blob into a text string and then passes it to the callback.