Function gloo_file::callbacks::read_as_array_buffer

pub fn read_as_array_buffer<F>(blob: &Blob, callback: F) -> FileReader where
    F: FnOnce(Result<ArrayBuffer, FileReadError>) + 'static, 

Asynchronously converts the blob into an array buffer and then passes it to the callback.

If the returned FileReader is dropped before the callback is called, the read will be cancelled.